requirements.txt
Flask
torch
torchvision
Pillow
paho-mqtt
